requirements.txt
# LLM Profiling Framework Dependencies

# 核心依赖
pyyaml>=6.0
numpy>=1.21.0
matplotlib>=3.5.0

# vLLM推理引擎 (需要CUDA环境)
# vllm>=0.2.0  # 请根据您的CUDA版本安装合适的vllm版本

# 可选依赖
pandas>=1.3.0  # 用于更高级的数据分析
plotly>=5.0.0  # 用于交互式可视化
seaborn>=0.11.0  # 用于更美观的统计图表

# 开发依赖
pytest>=7.0.0  # 测试框架
black>=22.0.0  # 代码格式化
flake8>=4.0.0  # 代码检查 